+++ packages/visualizations-react/stories/Chart/Formatting/Tooltips.stories.tsx
@@ -6,7 +6,7 @@ import { ChartSeriesType } from '@opendatasoft/visualizations';
 
 import { Chart } from 'src';
 import type { Props } from 'reactify';
-import { defaultSource } from '../../utils';
+import { defaultLinks } from '../../utils';
 
 enum Locales {
     English_UnitedStates = 'en-US',
@@ -164,7 +164,7 @@ export const Default: StoryObj<TemplateProps> = {
                 loading: false,
             },
             options: {
-                source: defaultSource,
+                links: defaultLinks,
                 labelColumn: 'x',
                 axis: {
                     x: {