+++ packages/visualizations-react/stories/Poi/PoiMap.stories.tsx
@@ -12,7 +12,7 @@ import type {
 } from '@opendatasoft/visualizations';
 
 import { PoiMap } from 'src';
-import { defaultSource, timeout } from '../utils';
+import { defaultLinks, timeout } from '../utils';
 
 import sources from './sources';
 
@@ -303,7 +303,7 @@ const options: PoiMapOptions = {
     title: 'Lorem Ipsum',
     subtitle: 'Dolor Sit Amet',
     description: 'More aria description',
-    sourceLink: defaultSource,
+    links: defaultLinks,
     images,
 };